Move Constants to Package
