+++ apps/bam.iobio/src/App.tsx
@@ -27,11 +27,13 @@ import {
 	IobioDataBroker,
 	IobioHistogram,
 	IobioPercentBox,
+	histogramKeys,
+	percentKeys,
 } from 'iobio-react-components';
 import type { BamContext, BamKey } from 'iobio-react-components/src/constants';
 import { useEffect, useState } from 'react';
 import './App.css';
-import { defaultBamContext, histogramKeys, iobioURL, isOutlierKey, percentKeys } from './util';
+import { defaultBamContext, iobioURL, isOutlierKey } from './util';
 
 const colors = ['red', 'orange', 'gold', 'aquamarine', 'cornflowerblue', 'purple'];
 const secondaryColors = ['lightpink', 'peachpuff', 'lemonchiffon', 'palegreen', 'powderblue', 'lavender'];
